% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doremifasolData_exemple.R
\name{doremifasolData_exemple}
\alias{doremifasolData_exemple}
\title{Récupérer le chemin vers les fichiers exemples de DoremifasolData}
\usage{
doremifasolData_exemple(fichier = NULL)
}
\arguments{
\item{fichier}{Nom du fichier. Si \code{NULL}, les fichiers exemples sont listés.}
}
\description{
{DoremifasolData} est livré avec un certain nombre de fichiers
(plats ou issus de tableurs) dans le répertoire \code{inst/extdata}
Cette fonction en facilite leur accès en retournant leurs chemins.
}
\examples{
doremifasolData_exemple()
doremifasolData_exemple("communes-01012019.csv")
}
